Use exact versions for packages, see description
By default when using `npm install --save`, npm supposes you will want
your dependencies to update with minor patches, however using this approach
makes the project difficult to maintain as every time someone does a npm install
and a new version is available, package-lock.josn is updated which then needs
to be commited alongside your real commit.

Also this makes the project subject to errors in updates to the packages used,
which makes it unstable. It is better to manually update a package when needed.