Add BuildRoadIndicator tests, explicit disable of defaultProps in Road
I preferred to disable line by line the defaultProps eslint warning, 
because I think it is a good idea to know that it is happening, but if 
the rule is not disabled, the branch coverage does not take the case 
when the default param is not set. So it is a tradeoff.